+++ animation_nodes/utils/depsgraph.py
@@ -3,5 +3,5 @@ import bpy
 def getActiveDepsgraph():
     return bpy.context.evaluated_depsgraph_get()
 
-def getEvaluatedID(id):
-    return id.evaluated_get(getActiveDepsgraph())
+def getEvaluatedID(idObject):
+    return idObject.evaluated_get(getActiveDepsgraph())