+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Fixed the name output in the Mesh Object Input node.
 - Fixed Slice List node for negative start index in Length mode.
 - Fixed duplication of the Curve Interpolation node.
+- Fixed the Ensure Animation Data option in the Mesh Object Output node.
 
 ### Changed